test(loading): migrate standalone
