+++ core/src/components/loading/test/standalone/loading.e2e.ts-snapshots/loading-standalone-diff-ios-ltr-Mobile-Chrome-linux.png
