+++ core/src/components/loading/test/standalone/loading.e2e.ts-snapshots/loading-standalone-diff-ios-rtl-Mobile-Chrome-linux.png
