+++ core/src/components/loading/test/standalone/loading.e2e.ts-snapshots/loading-standalone-diff-md-rtl-Mobile-Safari-linux.png
